requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
apscheduler==3.10.4
google-api-python-client==2.147.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
msal==1.31.0
httpx==0.27.2
discord.py==2.4.0
anthropic==0.39.0
